+++ content_script_all_frames.js
@@ -29,8 +29,6 @@ chrome.runtime.onMessage.addListener(function (request, sender, sendResponse) {
         if (isServiceNow) {
             addScript('/js/purify.min.js', false); //needed for safe html insertion required by FF
             addScript('inject.js', true);
-            // if (window.self === window.top)
-            //     addScript('admin/marker.js');
             if (document.getElementById("filter") != null || location.pathname.startsWith("/now/")) {
                 addScript('inject_parent.js');
             }
@@ -55,6 +53,9 @@ chrome.runtime.onMessage.addListener(function (request, sender, sendResponse) {
             }
             else if (location.pathname.startsWith("/merge_form_")) {
                 addScript('js/monaco/compare.js', false);
+            }
+            if (parent && parent.location.pathname.startsWith("/$studio.do")) {
+                addScript('/js/inject_studio.js', false);
             }