+++ publish.sh
@@ -43,6 +43,7 @@ sed -i '' "6s/.*/$var/" publish/manifest-onprem.json
 sed -i '' "6s/.*/$var/" publish/manifest-firefox-onprem.json
 sed -i '' "6s/.*/$var/" publish/manifest-edge.json
 
+
 #Remove the chrome extension and reunpack it
 rm -rf publish/chrome-snutils
 unzip publish/chrome-snutils.zip -d publish/chrome-snutils