+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # CHANGELOG.md
 
-## 7.3.4.2 (2024-03-01)
+## 7.4.0.3 (2024-03-05)
+Features:
+  - Set inital InitalTag to opacity 0.01 so only shows after settings applied
+  - Prevent open InitialTag shift-click opening in new window
+
+## 7.4.0.0/1 (2024-03-05)
+Features:
+  - Beta release of InstanceTag. More details to follow. Early adopters can enable it in the settings tab of the popup.
+  - 7.4.0.1 adds a fix for checkbox setting in the settings tab
+
+## 7.3.4.4 (2024-03-01)
 Fixes / changes:
   - Adjusted snuHyperlinkifyWorkNotes function to prevent issues when [code] tags are used in worknotes
   - Preparations for Instance Button feature, that will be added in an upcoming release